requirements.txt
groq
python-dotenv
streamlit
pillow